Issue with missing Seconds of Week in IGS R03 sp3 orbits
The seconds of the week field in the second line of the IGSR03 combined repro3 SP3 orbits

Columns 9-23        Seconds of Week  e.g. 518400.00000000

are all declared as 0.00000000 (at least in all the files we have checked so far). This makes software such as GPSPACE not work, since variable SECOW is read from the header and used in subsequent computations. 

Here is one example from a random day:

#dP2010  1  2  0  0  0.00000000     288 ORBIT IGSR3 HLM IGS
## 1564      0.00000000   300.00000000 55198 0.0000000000000

while the repro2 orbits show

#cP2010  1  2  0  0  0.00000000      96 ORBIT IGb08 HLM  IGS
## 1564 518400.00000000   900.00000000 55198 0.0000000000000

I have modified GPSPACE to check if SECOW == 0 and, if true, recompute it using the date of the ephemeris file. Note that some assumptions were made, such as that the first epoch in the sp3 file is 0h 0m 0s. This should provide a temporary solution until the SP3 IGS R03 files are fixed.